+++ finject_generator/lib/src/builders/annotation_analize/finject_analizer.dart
@@ -10,6 +10,7 @@ class Analizer {
 
     _attachConstructorInjection(injectionDefinition, classInfo.constructors);
     _attachFieldsInjections(injectionDefinition, classInfo.fields);
+    _attachSuperClassInjections(injectionDefinition, classInfo);
     _attachMethodsInjection(injectionDefinition, classInfo.methods);
     injectionDefinition.prunRedundantInjections();
     injectionDefinition.margeDependencies();
@@ -20,7 +21,7 @@ class Analizer {
       InjectorDs injection, List<ConstructorElement> constructors) {
     _validateInjectConstructor(constructors);
     for (var element in constructors) {
-      MethodInjection constructorInjection =
+      var constructorInjection =
           _processParemetersOfCostructor(element.parameters);
       injection.constructorInjection = constructorInjection;
       break;
@@ -30,8 +31,8 @@ class Analizer {
   MethodInjection _processParemetersOfCostructor(
       List<ParameterElement> parameters) {
     var constructorInjection = MethodInjection();
-    for (ParameterElement element in parameters) {
-      ClassElement classInfo = getType(element.type);
+    for (var element in parameters) {
+      var classInfo = getType(element.type);
       if (element.isPositional) {
         constructorInjection.addOrderedParameter(
             convert(classInfo), findName(element.metadata));
@@ -50,9 +51,9 @@ class Analizer {
   }
 
   void _validateInjectConstructor(List<ConstructorElement> constructors) {
-    int numberOfInjection = 0;
-    for (ConstructorElement element in constructors) {
-      if (hasAnnotation(element.metadata, "Inject")) {
+    var numberOfInjection = 0;
+    for (var element in constructors) {
+      if (hasAnnotation(element.metadata, 'Inject')) {
         numberOfInjection++;
       }
       if (element.isDefaultConstructor) {
@@ -69,13 +70,31 @@ class Analizer {
 
   void _attachFieldsInjections(
       InjectorDs injection, List<FieldElement> fields) {
-    for (FieldElement element in fields) {
-      if (hasAnnotation(element.metadata, "Inject")) {
-        ClassElement classInfo = element.type.element as ClassElement;
+    for (var element in fields) {
+      if (hasAnnotation(element.metadata, 'Inject')) {
+        var classInfo = element.type.element as ClassElement;
         injection.fieldInjection.addNamedParameter(
             element.name, convert(classInfo), findName(element.metadata));
       }
     }
+  }
+
+  void _attachSuperClassInjections(InjectorDs injection, ClassElement element) {
+    if (element.supertype == null) {
+      return;
+    }
+
+    var superTypeElement = element.supertype.element;
+
+    for (var element in superTypeElement.fields) {
+      if (hasAnnotation(element.metadata, 'Inject')) {
+        var classInfo = element.type.element as ClassElement;
+        injection.fieldInjection.addNamedParameter(
+            element.name, convert(classInfo), findName(element.metadata));
+      }
+    }
+
+    _attachSuperClassInjections(injection, superTypeElement);
   }
 
   void _attachMethodsInjection(