+++ finject_generator/lib/src/builders/annotation_analize/finject_configuration_analizer.dart
@@ -6,46 +6,50 @@ import 'finject_analizer.dart';
 
 class FinjectConfigurationAnalizer extends Analizer {
   Iterable<InjectorDs> analize(Element element) {
-    List<InjectorDs> injections = [];
+    var injections = <InjectorDs>[];
     if (element is ClassElement) {
-      ClassElement classInfo = element;
-      InjectorDs injectionDefinition = prepareInjectionDefinition(classInfo);
+      var classInfo = element;
+      var injectionDefinition = prepareInjectionDefinition(classInfo);
 
       injectionDefinition.singleton = true;
 
       injections.add(injectionDefinition);
 
-      for (MethodElement method in classInfo.methods) {
+      for (var method in classInfo.methods) {
         if (method.isPrivate) {
           continue;
         }
 
         injectionDefinition = InjectorDs();
-        injectionDefinition.typeName = convert(method.returnType.element as ClassElement);
+        injectionDefinition.typeName =
+            convert(method.returnType.element as ClassElement);
         injectionDefinition.factoryTypeName = convert(classInfo);
 
         attachFactoryMethod(injectionDefinition, method);
 
         injectionDefinition.singleton =
-            hasAnnotation(method.metadata, "Singleton");
+            hasAnnotation(method.metadata, 'Singleton');
 
-        for (ElementAnnotation annotation in method.metadata) {
-          ConstructorElement annotationInfo = annotation.element as ConstructorElement;
-          ClassElement annotationType = annotationInfo.enclosingElement;
+        for (var annotation in method.metadata) {
+          var annotationInfo = annotation.element as ConstructorElement;
+          var annotationType = annotationInfo.enclosingElement;
 
-          if (annotationType.name == "Scoped") {
+          if (annotationType.name == 'Scoped') {
             var result = annotation.computeConstantValue();
             injectionDefinition.scopeName =
-                result.getField("name").toStringValue();
+                result.getField('name').toStringValue();
           }
-          if (annotationType.name == "Named") {
+          if (annotationType.name == 'Named') {
             var result = annotation.computeConstantValue();
-            injectionDefinition.name = result.getField("name").toStringValue();
+            injectionDefinition.name = result.getField('name').toStringValue();
           }
           if (annotationType.name == "Profile") {
             var result = annotation.computeConstantValue();
-            injectionDefinition.profiles =
-                result.getField("profileNames").toListValue().map((value) => value.toStringValue()).toSet();
+            injectionDefinition.profiles = result
+                .getField("profileNames")
+                .toListValue()
+                .map((value) => value.toStringValue())
+                .toSet();
           }
         }
 
@@ -59,8 +63,8 @@ class FinjectConfigurationAnalizer extends Analizer {
   void attachFactoryMethod(InjectorDs injection, MethodElement method) {
     var methodInjection = MethodInjection();
     methodInjection.name = method.name;
-    for (ParameterElement element in method.parameters) {
-      ClassElement classInfo = getType(element.type);
+    for (var element in method.parameters) {
+      var classInfo = getType(element.type);
       if (element.isPositional) {
         methodInjection.addOrderedParameter(
             convert(classInfo), findName(method.metadata));