+++ finject_generator/lib/src/builders/annotation_analize/finject_injectable_analizer.dart
@@ -6,26 +6,26 @@ import 'finject_analizer.dart';
 
 class FinjectInjectableAnalizer extends Analizer {
   Iterable<InjectorDs> analize(Element element) {
-    List<InjectorDs> injections = [];
+    var injections = <InjectorDs>[];
 
     if (element is ClassElement) {
-      ClassElement classInfo = element;
-      InjectorDs injectionDefinition = prepareInjectionDefinition(classInfo);
+      var classInfo = element;
+      var injectionDefinition = prepareInjectionDefinition(classInfo);
 
       injectionDefinition.singleton =
-          hasAnnotation(classInfo.metadata, "Singleton");
+          hasAnnotation(classInfo.metadata, 'Singleton');
 
-      for (ElementAnnotation annotation in classInfo.metadata) {
-        ConstructorElement annotationInfo = annotation.element as ConstructorElement;
-        ClassElement annotationType = annotationInfo.enclosingElement;
-        if (annotationType.name == "Scoped") {
+      for (var annotation in classInfo.metadata) {
+        var annotationInfo = annotation.element as ConstructorElement;
+        var annotationType = annotationInfo.enclosingElement;
+        if (annotationType.name == 'Scoped') {
           var result = annotation.computeConstantValue();
           injectionDefinition.scopeName =
-              result.getField("name").toStringValue();
+              result.getField('name').toStringValue();
         }
-        if (annotationType.name == "Named") {
+        if (annotationType.name == 'Named') {
           var result = annotation.computeConstantValue();
-          injectionDefinition.name = result.getField("name").toStringValue();
+          injectionDefinition.name = result.getField('name').toStringValue();
         }
       }