+++ finject_generator/lib/src/builders/code_info_extraction.dart
@@ -6,12 +6,12 @@ Map<String, int> knownLibraries = {};
 int currentLibraryNumber = 0;
 
 String findName(List<ElementAnnotation> annotations) {
-  for (ElementAnnotation annotation in annotations) {
-    ConstructorElement annotationInfo = annotation.element as ConstructorElement;
-    ClassElement annotationType = annotationInfo.enclosingElement;
-    if (annotationType.name == "Named") {
+  for (var annotation in annotations) {
+    var annotationInfo = annotation.element as ConstructorElement;
+    var annotationType = annotationInfo.enclosingElement;
+    if (annotationType.name == 'Named') {
       var result = annotation.computeConstantValue();
-      return result.getField("name").toStringValue();
+      return result.getField('name').toStringValue();
     }
   }
   return null;
@@ -21,8 +21,8 @@ TypeInfo convert(ClassElement element) {
   if (element == null) {
     throw InjectorValidationError();
   }
-  Uri uriOfClass = element.librarySource.uri;
-  int libraryId = 0;
+  var uriOfClass = element.librarySource.uri;
+  var libraryId = 0;
   if (!knownLibraries.containsKey(uriOfClass.path)) {
     currentLibraryNumber++;
     knownLibraries[uriOfClass.path] = currentLibraryNumber;
@@ -30,7 +30,7 @@ TypeInfo convert(ClassElement element) {
   libraryId = knownLibraries[uriOfClass.path];
 
   return TypeInfo(
-      uriOfClass.scheme, uriOfClass.path, element.name, "id$libraryId");
+      uriOfClass.scheme, uriOfClass.path, element.name, 'id$libraryId');
 }
 
 ClassElement getType(DartType type) {
@@ -42,9 +42,9 @@ ClassElement getType(DartType type) {
 }
 
 bool hasAnnotation(List<ElementAnnotation> metadata, String type) {
-  for (ElementAnnotation annotation in metadata) {
-    ConstructorElement annotationInfo = annotation.element as ConstructorElement;
-    ClassElement classInfo = annotationInfo.enclosingElement;
+  for (var annotation in metadata) {
+    var annotationInfo = annotation.element as ConstructorElement;
+    var classInfo = annotationInfo.enclosingElement;
     if (classInfo.name == type) {
       return true;
     }