+++ finject_generator/lib/src/builders/di_code_builder.dart
@@ -30,8 +30,8 @@ class DiCodeBuilder implements Builder {
     };
   }
 
-  bool notContainsOneOf(Set<String> injectorProfiles,
-      List<String> activeProfiles) {
+  bool notContainsOneOf(
+      Set<String> injectorProfiles, List<String> activeProfiles) {
     for (String injectorProfile in injectorProfiles) {
       if (activeProfiles.contains(injectorProfile)) {
         return false;
@@ -42,10 +42,10 @@ class DiCodeBuilder implements Builder {
 
   @override
   Future<void> build(BuildStep buildStep) async {
-    List<ClassSpec> allInjectors = [];
-    Map<String, List<InjectorDs>> scopes = {};
-    List<InjectorDs> allTypes = [];
-    List<String> profiles = [];
+    var allInjectors = <ClassSpec>[];
+    var scopes = <String, List<InjectorDs>>{};
+    var allTypes = <InjectorDs>[];
+    var profiles = <String>[];
 
     await for (final input in buildStep.findAssets(declaratedProfiles)) {
       var library = await buildStep.resolver.libraryFor(input);
@@ -54,7 +54,7 @@ class DiCodeBuilder implements Builder {
           .where(
               (TopLevelVariableElement element) => element.name == "profiles")
           .map((TopLevelVariableElement element) =>
-          element.computeConstantValue())
+              element.computeConstantValue())
           .toList();
 
       if (profilesValue.length == 1) {
@@ -66,21 +66,21 @@ class DiCodeBuilder implements Builder {
     }
 
     await for (final input in buildStep.findAssets(_allFilesInLib)) {
-      if (input.path.endsWith("summary.json")) {
-        String injectorJson = await buildStep.readAsString(input);
+      if (input.path.endsWith('summary.json')) {
+        var injectorJson = await buildStep.readAsString(input);
 
-        List<InjectorDs> readData = (jsonDecode(injectorJson) as Iterable)
+        var readData = (jsonDecode(injectorJson) as Iterable)
             .map((value) => InjectorDs.fromJson(value as Map<String, dynamic>))
             .toList();
 
-        for (InjectorDs oneInjectable in readData) {
+        for (var oneInjectable in readData) {
           if (oneInjectable.profiles.isNotEmpty &&
               notContainsOneOf(oneInjectable.profiles, profiles)) {
             continue;
           }
 
           if (oneInjectable.scopeName != null) {
-            List<InjectorDs> scopeList = scopes[oneInjectable.scopeName];
+            var scopeList = scopes[oneInjectable.scopeName];
             if (scopeList == null) {
               scopeList = <InjectorDs>[];
               scopes[oneInjectable.scopeName] = scopeList;
@@ -97,78 +97,74 @@ class DiCodeBuilder implements Builder {
 
     allInjectors.add(createClassesForScope(scopes));
 
-    FileSpec fileSpec = FileSpec.build(
+    var fileSpec = FileSpec.build(
         dependencies: [...generateImport(allTypes)],
         classes: allInjectors,
         properties: [
           PropertySpec.ofMapByToken(
-            "injectorMapper",
+            'injectorMapper',
             keyType: TypeToken.of(Qualifier),
-            valueType: TypeToken.ofFullName("Injector"),
+            valueType: TypeToken.ofFullName('Injector'),
             defaultValue: {},
           ),
           PropertySpec.ofMapByToken(
-            "factoryMapper",
+            'factoryMapper',
             keyType: TypeToken.of(Qualifier),
-            valueType: TypeToken.ofFullName("Factory"),
+            valueType: TypeToken.ofFullName('Factory'),
             defaultValue: {},
           ),
         ],
         methods: [
-          MethodSpec.build("init", codeBlock: generateCodeForMappers(allTypes))
+          MethodSpec.build('init', codeBlock: generateCodeForMappers(allTypes))
         ]);
 
-    DartFile dartFile = DartFile.fromFileSpec(fileSpec);
+    var dartFile = DartFile.fromFileSpec(fileSpec);
 
     final output = _allFileOutput(buildStep);
     return buildStep.writeAsString(output, dartFile.outputContent());
   }
 
   ClassSpec createClassesForScope(Map<String, List<InjectorDs>> scopes) {
-    return ClassSpec.build("ScopeFactoryImpl",
+    return ClassSpec.build('ScopeFactoryImpl',
         superClass: TypeToken.of(ScopeFactory),
         methods: [
-          MethodSpec.build("createScope",
+          MethodSpec.build('createScope',
               returnType: TypeToken.of(Scope),
               parameters: [
-                ParameterSpec.build("scopeName", type: TypeToken.ofString())
+                ParameterSpec.build('scopeName', type: TypeToken.ofString())
               ],
               codeBlock:
-              CodeBlockSpec.lines(generateMethodCodeForScopes(scopes)))
+                  CodeBlockSpec.lines(generateMethodCodeForScopes(scopes)))
         ]);
   }
 }
 
 List<String> generateMethodCodeForScopes(Map<String, List<InjectorDs>> scopes) {
-  List<String> linesOfCode = [];
+  var linesOfCode = <String>[];
 
-  for (String scope in scopes.keys) {
+  for (var scope in scopes.keys) {
     linesOfCode.add("if (scopeName == '$scope') {");
-    linesOfCode.add("return Scope([");
-    for (InjectorDs injectable in scopes[scope]) {
+    linesOfCode.add('return Scope([');
+    for (var injectable in scopes[scope]) {
       linesOfCode.add(
-          "ScopeEntry<Injector>(const ${generateKeyForInjector(
-              injectable)}, ${generatePrefixClassName(
-              injectable)}_Injector()),");
+          'ScopeEntry<Injector>(const ${generateKeyForInjector(injectable)}, ${generatePrefixClassName(injectable)}_Injector()),');
       linesOfCode.add(
-          "ScopeEntry<Factory>(const ${generateKeyForInjector(
-              injectable)}, ${generatePrefixClassName(
-              injectable)}_Factory()),");
+          'ScopeEntry<Factory>(const ${generateKeyForInjector(injectable)}, ${generatePrefixClassName(injectable)}_Factory()),');
     }
-    linesOfCode.add("]);");
-    linesOfCode.add("}");
+    linesOfCode.add(']);');
+    linesOfCode.add('}');
   }
 
-  linesOfCode.add("//default value");
-  linesOfCode.add("return null;");
+  linesOfCode.add('//default value');
+  linesOfCode.add('return null;');
   return linesOfCode;
 }
 
 Iterable<ClassSpec> createClassesForSummary(InjectorDs readData) sync* {
   yield ClassSpec.build(
-    generatePrefixClassName(readData) + "_Injector",
+    generatePrefixClassName(readData) + '_Injector',
     superClass: TypeToken.ofFullName(
-      "Injector<${generateTypeFromTypeInfo(readData.typeName)}>",
+      'Injector<${generateTypeFromTypeInfo(readData.typeName)}>',
     ),
     doc: DocSpec.text(
         'this is injector for ' + readData.typeName.className + ' class'),
@@ -178,13 +174,13 @@ Iterable<ClassSpec> createClassesForSummary(InjectorDs readData) sync* {
         'inject',
         parameters: [
           ParameterSpec.build(
-            "instance",
+            'instance',
             type: TypeToken.ofName2(
               generateTypeFromTypeInfo(readData.typeName),
             ),
           ),
-          ParameterSpec.normal("injectionProvider",
-              type: TypeToken.ofFullName("InjectionProvider")),
+          ParameterSpec.normal('injectionProvider',
+              type: TypeToken.ofFullName('InjectionProvider')),
         ],
         returnType: TypeToken.ofVoid(),
         codeBlock: generateCodeForInjector(readData),
@@ -192,22 +188,22 @@ Iterable<ClassSpec> createClassesForSummary(InjectorDs readData) sync* {
     ],
   );
   yield ClassSpec.build(
-    generatePrefixClassName(readData) + "_Factory",
+    generatePrefixClassName(readData) + '_Factory',
 //          metas: [MetaSpec.of('Object()')],
     doc: DocSpec.text('this is factory for ' +
         generateTypeFromTypeInfo(readData.typeName) +
         ' class'),
     superClass: TypeToken.ofFullName(
-        "Factory<${generateTypeFromTypeInfo(readData.typeName)}>"),
+        'Factory<${generateTypeFromTypeInfo(readData.typeName)}>'),
     properties: createPropertyFactoryForScopedSingleton(readData).toList(),
     methods: [
       MethodSpec.build(
         'create',
         returnType:
-        TypeToken.ofName2(generateTypeFromTypeInfo(readData.typeName)),
+            TypeToken.ofName2(generateTypeFromTypeInfo(readData.typeName)),
         parameters: [
-          ParameterSpec.normal("injectionProvider",
-              type: TypeToken.ofFullName("InjectionProvider")),
+          ParameterSpec.normal('injectionProvider',
+              type: TypeToken.ofFullName('InjectionProvider')),
         ],
         codeBlock: generateCodeForFactory(readData),
       ),
@@ -218,7 +214,7 @@ Iterable<ClassSpec> createClassesForSummary(InjectorDs readData) sync* {
 Iterable<PropertySpec> createPropertyFactoryForScopedSingleton(
     InjectorDs data) sync* {
   if (data.singleton) {
-    yield PropertySpec.of("cache",
+    yield PropertySpec.of('cache',
         type: TypeToken.ofName2(generateTypeFromTypeInfo(data.typeName)));
   }
 }
@@ -226,7 +222,7 @@ Iterable<PropertySpec> createPropertyFactoryForScopedSingleton(
 Iterable<PropertySpec> createPropertyInjectorForScopedSingleton(
     InjectorDs data) sync* {
   if (data.singleton) {
-    yield PropertySpec.of("cache",
+    yield PropertySpec.of('cache',
         type: TypeToken.ofBool(), defaultValue: false);
   }
 }