+++ finject_generator/lib/src/builders/summary_builder.dart
@@ -10,13 +10,13 @@ class BuildSummary extends Builder {
   @override
   Map<String, List<String>> get buildExtensions {
     return {
-      ".dart": [_extension],
+      '.dart': [_extension],
     };
   }
 
   @override
   Future<void> build(BuildStep buildStep) async {
-    AssetId output = buildStep.inputId.changeExtension(_extension);
+    var output = buildStep.inputId.changeExtension(_extension);
     final lib = await buildStep.inputLibrary;
 
     var createdUnit =