+++ finject_generator/lib/src/dart_poet_extensions/dependency_spec_ext.dart
@@ -7,7 +7,11 @@ class DependencySpecExt extends DependencySpec {
 
   @override
   String code({Map<String, dynamic> args = const {}}) {
-    String result = super.code(args: args).replaceAll(";", "");
-    return "$result as $libraryId;";
+    if (!route.contains('dart:core')) {
+      var result = super.code(args: args).replaceAll(';', '');
+      return '$result as $libraryId;';
+    } else {
+      return super.code(args: args);
+    }
   }
 }